moved DTO model and DTO Mappers to own project and domain, respectively
