+++ ClientManagementService/ClientManagementService.DTO/PasswordDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace ClientManagementService.API.DTO
+﻿namespace ClientManagementService.DTO
 {
     public class PasswordDTO
     {
