+++ ClientManagementService/ClientManagementService.DTO/PetTypeDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace ClientManagementService.API.DTO
+﻿namespace ClientManagementService.DTO
 {
     public class PetTypeDTO
     {
