+++ ClientManagementService/ClientManagementService.DTO/VaccineDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace ClientManagementService.API.DTO
+﻿namespace ClientManagementService.DTO
 {
     public class VaccineDTO
     {
