+++ ClientManagementService/ClientManagementService.Domain/ClientManagementService.Domain.csproj
@@ -9,6 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\ClientManagementService.DTO\ClientManagementService.DTO.csproj" />
     <ProjectReference Include="..\ClientManagementService.Infrastructure\ClientManagementService.Infrastructure.csproj" />
   </ItemGroup>
